# Conflicts:
#	src/main/resources/db/changelog/changes/003-modified-schema.sql

+++ src/main/java/com/williamfeliciano/sba_grocerystore/dtos/DetailedItemResponseDto.java
@@ -18,6 +18,6 @@ public class DetailedItemResponseDto {
     private String price;
     private String category;
     private String pictureUrl;
-    private int weight;
+    private double weight;
     private int stockAmount;
 }
